Instruction class usage refactoring
